backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python-headless>=4.8.0
easyocr>=1.7.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
celery>=5.3.0
redis>=5.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
